MODLD-829: Create UpdatedDateMapper
